move routed components to views folder
