+++ src/router.js
@@ -3,28 +3,28 @@ import VueResource from 'vue-resource'
 import VueRouter from 'vue-router'
 import VueSocketio from 'vue-socket.io'
 
-import Contact from './components/Contact'
-import Legal from './components/Legal'
-import Logout from './components/Logout'
-import LoginForm from './components/LoginForm'
-import Registration from './components/Registration'
-import ResetPasswordForm from './components/ResetPasswordForm'
-import SetPasswordForm from './components/SetPasswordForm'
-import Onboarding from './components/Onboarding'
-import Dashboard from './components/Dashboard'
-import Session from './components/Session'
-import Action from './components/Action'
-import Schedule from './components/Schedule'
-import Resources from './components/Resources'
-import Feedback from './components/Feedback'
-import Training from './components/Training'
-import Quiz from './components/Quiz'
-import Review from './components/Review'
-import Profile from './components/Profile'
-import Calendar from './components/Calendar'
-import SubmitQuestion from './components/views/SubmitQuestion'
-import Inbox from './components/views/Inbox'
-import SendAnswer from './components/views/SendAnswer'
+import Contact from './views/Contact'
+import Legal from './views/Legal'
+import Logout from './views/Logout'
+import LoginForm from './views/LoginForm'
+import Registration from './views/Registration'
+import ResetPasswordForm from './views/ResetPasswordForm'
+import SetPasswordForm from './views/SetPasswordForm'
+import Onboarding from './views/Onboarding'
+import Dashboard from './views/Dashboard'
+import Session from './views/Session'
+import Action from './views/Action'
+import Schedule from './views/Schedule'
+import Resources from './views/Resources'
+import Feedback from './views/Feedback'
+import Training from './views/Training'
+import Quiz from './views/Quiz'
+import Review from './views/Review'
+import Profile from './views/Profile'
+import Calendar from './views/Calendar'
+import SubmitQuestion from './views/SubmitQuestion'
+import Inbox from './views/Inbox'
+import SendAnswer from './views/SendAnswer'
 
 import AuthService from './services/AuthService'
 import OnboardingService from './services/OnboardingService'
