+++ src/views/Contact.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import UserService from 'src/services/UserService'
+import UserService from '../services/UserService'
 
 export default {
   data () {
