+++ src/views/Feedback.vue
@@ -90,7 +90,7 @@
 </template>
 
 <script>
-import UserService from 'src/services/UserService'
+import UserService from '../services/UserService'
 import NetworkService from '../services/NetworkService'
 
 export default {
