+++ src/views/Profile.vue
@@ -125,7 +125,7 @@
 </template>
 
 <script>
-import UserService from 'src/services/UserService'
+import UserService from '../services/UserService'
 
 export default {
   data () {
