+++ src/views/Resources.vue
@@ -153,7 +153,7 @@
 </template>
 
 <script>
-import UserService from 'src/services/UserService'
+import UserService from '../services/UserService'
 
 export default {
   data () {
