+++ src/views/SendAnswer.vue
@@ -33,13 +33,13 @@
 </template>
 
 <script>
-import UserService from '../../services/UserService'
-import DownloadService from '../../services/DownloadService'
-import StudentQuestionService from '../../services/StudentQuestionService'
+import UserService from '../services/UserService'
+import DownloadService from '../services/DownloadService'
+import StudentQuestionService from '../services/StudentQuestionService'
 
-import BasicTemplate from '../organisms/BasicTemplate'
-import MessageForm from '../organisms/MessageForm'
-import Modal from '../molecules/Modal'
+import BasicTemplate from '../components/organisms/BasicTemplate'
+import MessageForm from '../components/organisms/MessageForm'
+import Modal from '../components/molecules/Modal'
 
 export default {
   components: {
@@ -136,7 +136,7 @@ export default {
   transform: translate(-50%, -50%);
   width: 24px;
   height: 24px;
-  background-image: url('../../assets/attachment_icon.svg');
+  background-image: url('../assets/attachment_icon.svg');
 }
 .attachment-list__content {
   padding: 20px;
